% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultimate_deconvolution.R
\name{dmixlike}
\alias{dmixlike}
\title{Calculates the gaussian mixture density.}
\usage{
dmixlike(x_mat, pi_vec, v_mat, s_mat, log = FALSE)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations and
the columns index the variables.}

\item{pi_vec}{A vector of mixture proportions.}

\item{v_mat}{A matrix of square-roots of the rank-1 portion of the
the mixing covariance matrices. The columns index the mixing
components, the rows index the variables.}

\item{s_mat}{A matrix of diagonal elements of the diagonal portion
of the mixing covariance matrices. The columns index the
variables and the rows index observations.}

\item{log}{A logical. Should we return the log-density
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
This function assumes that the mixing means are all zeros and
the mixing covariances are structured as a rank-1 matrix plus
a diagonal matrix.
}
\author{
David Gerard
}
