% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmixlike_cpp}
\alias{dmixlike_cpp}
\title{Calculates the gaussian mixture density.}
\usage{
dmixlike_cpp(x_mat, s_mat, v_mat, pi_vec, return_log = FALSE)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations and
the columns index the variables.}

\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of square-roots of the rank-1 mixing
covariance matrices. The columns index the mixing
components, the rows index the variables.}

\item{pi_vec}{A vector of mixture proportions.}

\item{return_log}{A logical. Should we return the log-density
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
This function assumes that the mixing means are all zeros and
the mixing covariances are rank-1 matrices. Each observation
has its own independent noise (variances collected in
\code{s_mat}).
}
\author{
David Gerard
}
