% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnorm_rank1}
\alias{dnorm_rank1}
\title{Multivariate normal density at mean 0 and a covariance that is the sum
of a rank-1 matrix and a diagonal matrix.}
\usage{
dnorm_rank1(x, v, s_diag, mu, return_log = FALSE)
}
\arguments{
\item{x}{The data.}

\item{v}{The square-root of the rank-1 matrix.}

\item{s_diag}{The diagonal elements of the diagonal component of the
covariance matrix.}

\item{mu}{The mean. Defaults to 0.}

\item{return_log}{Should we return the log-density (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Multivariate normal density at mean 0 and a covariance that is the sum
of a rank-1 matrix and a diagonal matrix.
}
\author{
David Gerard
}
