% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultimate_deconvolution.R
\name{dr1_norm}
\alias{dr1_norm}
\title{Multivariate normal density at mean 0 and a
covariance that is the sum of a rank-1 matrix and a diagonal matrix.}
\usage{
dr1_norm(x, v, s_diag, mu = rep(0, length(x)), log = FALSE)
}
\arguments{
\item{x}{The data.}

\item{v}{The square-root of the rank-1 matrix.}

\item{s_diag}{The diagonal elements of the diagonal component of the
covariance matrix.}

\item{mu}{The mean. Defaults to 0.}

\item{log}{A logical. Should we return the log-density (\code{TRUE}) or the
density (\code{FALSE})?}
}
\description{
Multivariate normal density at mean 0 and a
covariance that is the sum of a rank-1 matrix and a diagonal matrix.
}
\author{
David Gerard
}
