% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_cpp}
\alias{em_cpp}
\title{C++ version of EM algorithm.}
\usage{
em_cpp(x_mat, s_mat, v_mat, pi_vec, itermax = 500L, tol = 10^-5,
  plot_iter = FALSE)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations
and the columns index the variables.}

\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of initial values for the low-rank mixture
covariances.}

\item{pi_vec}{A vector of initial values of the mixing proportions.}

\item{itermax}{The maximum number of fixed-point iterations for
the EM to run.}

\item{tol}{The tolerance for the stopping criterion. The current
stopping criterion is the ratio of successive
likelihoods minus 1.}

\item{plot_iter}{A logical. Should we plot updates (\code{TRUE}) or not (\code{FALSE})?}
}
\value{
A list of the following elements:

    \code{pi_vec}: The final estimate of the mixing proportions.

    \code{v_mat}: The final estimate of the square roots of the rank-1 covariance matrices.
        These are the factors.

    \code{llike_vec}: The vector of log-likelihoods. Should be increasing.

    \code{convergence}: A value of \code{0} indicates convergence. A value of \code{1} indicates that
        the limit \code{itermax} has been reached. A vlue of \code{2} indicates that the user
        interupted the optimization.

    \code{loadings}: An estimate for the loadings.

    \code{w_mat}: An estimate for the probability of being in a group.
}
\description{
C++ version of EM algorithm.
}
\author{
David Gerard
}
