% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_fix}
\alias{em_fix}
\title{Fixed iteration from the EM algorithm.}
\usage{
em_fix(x_mat, s_mat, v_mat, pi_vec)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations and
the columns index the variables.}

\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of square-roots of the rank-1 mixing
covariance matrices. The columns index the mixing
components, the rows index the variables.}

\item{pi_vec}{A vector of mixture proportions.}
}
\description{
Fixed iteration from the EM algorithm.
}
\author{
David Gerard
}
