% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_r}
\alias{em_r}
\title{R version of the EM algorithm.}
\usage{
em_r(x_mat, s_mat, v_mat, pi_vec, itermax = 500, tol = 10^-5,
  plot_iter = FALSE)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations
and the columns index the variables.}

\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of initial values for the low-rank mixture
covariances.}

\item{pi_vec}{A vector of initial values of the mixing proportions.}

\item{itermax}{The maximum number of fixed-point iterations for
the EM to run.}

\item{tol}{The tolerance for the stopping criterion. The current
stopping criterion is the ratio of successive
likelihoods minus 1.}

\item{plot_iter}{A logical. Should we plot the likelihood at
each step (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
R version of the EM algorithm.
}
\author{
David Gerard
}
