% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_llike}
\alias{plot_llike}
\title{Basic line plot for \code{\link{em_cpp}}.}
\usage{
plot_llike(llike_vec, itermax)
}
\arguments{
\item{llike_vec}{The y-values}

\item{itermax}{A positive integer. The x-values are \code{1:itermax}.}
}
\description{
Basic line plot for \code{\link{em_cpp}}.
}
\author{
David Gerard
}
