% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{rmixmultnorm}
\alias{rmixmultnorm}
\title{Random draw from mixture of multivariate normals.}
\usage{
rmixmultnorm(s_mat, v_mat, pi_vec)
}
\arguments{
\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of square-roots of the rank-1 mixing
covariance matrices. The columns index the mixing
components, the rows index the variables.}

\item{pi_vec}{A vector of mixture proportions.}
}
\description{
It is assumed that the mixing covariances are rank-1 (with their square roots in
\code{v_mat}) and the observation covariances are diagonal (with their variances in s_mat).
}
\details{
The number of observations drawn is equal to \code{nrow(s_mat)}.
}
\seealso{
dmixlike
}
\author{
David Gerard
}
