% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{ultimate_deconvolution}
\alias{ultimate_deconvolution}
\title{EM algorithm to fit multivariate Guassian convolution problem.}
\usage{
ultimate_deconvolution(x_mat, s_mat, v_mat, pi_vec, itermax = 500,
  tol = 10^-5, print_iter = FALSE)
}
\arguments{
\item{x_mat}{A matrix of data. The rows index the observations
and the columns index the variables.}

\item{s_mat}{A matrix of variances (NOT standard deviations).
The rows index the observations and the columns index the
variables.}

\item{v_mat}{A matrix of initial values for the low-rank mixture
covariances.}

\item{pi_vec}{A vector of initial values of the mixing proportions.}

\item{itermax}{The maximum number of fixed-point iterations for
the EM to run.}

\item{tol}{The tolerance for the stopping criterion. The current
stopping criterion is the ratio of successive
likelihoods minus 1.}

\item{print_iter}{A logical. Should we print the likelihood at
each step (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
EM algorithm to fit multivariate Guassian convolution problem.
}
\author{
David Gerard
}
