% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UltimateDeconvolution.R
\docType{package}
\name{UltimateDeconvolution}
\alias{UltimateDeconvolution}
\alias{UltimateDeconvolution-package}
\title{UltimateDeconvolution: Rank-1 deconvolution problem.}
\description{
UltimateDeconvolution: Rank-1 deconvolution problem.
}
\author{
David Gerard
}
